requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==2.3.2
Flask-Cors==4.0.0
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-vision==3.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
urllib3==2.2.0
Werkzeug==3.0.1
